feat: storybook improvements, move the stories into the components
