+++ packages/storybook/src/components/atoms/LoaderIcon/loader.stories.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
 
-import { types } from '../../../src/components/atoms/Button/props';
-import LoaderIcon from '../../../src/components/atoms/LoaderIcon';
-import { sizes } from '../../../src/components/atoms/LoaderIcon/props';
+import { types } from '../Button/props';
+
+import { sizes } from './props';
+
+import LoaderIcon from '.';
 
 export default {
   component: LoaderIcon,
