--- packages/storybook/stories/titlePath.js
@@ -1,9 +1,0 @@
-export const projects = {
-  template: 'template',
-};
-
-export const componentTypes = {
-  atom: 'atom',
-  molecule: 'molecule',
-  icon: 'icon',
-};
